Tried to used 3 versions of SMOTE
I tried to use 3 version of SMOTE: SMOTE, SMOTEN and SMOTENC, all returned an error. Sent a Whatsapp with screen shot of the error. We'll need to look in to that before we move ahead. That or we'll need to encode all the categorcal data, int which case a tree model may not be suitable.